Added: Keyboard handling to main window. Other: Minor cleanup and few small fixes.
git-svn-id: svn://svn.CzDC.org/PtokaX@53 4c629181-62e4-40b4-9e0d-b9f719119ca2

+++ core/DcCommands.cpp
@@ -1149,7 +1149,7 @@ void cDcCommands::ConnectToMe(User * curUser, char * sData, const uint32_t &iLen
     curUser->ui32BoolBits |= User::BIT_ACTIVE;
 
     // full data only and allow blocking
-	if(ScriptManager->Arrival(curUser, sData, iLen, bMulti == false ? ScriptMan::CONNECTTOME_ARRIVAL : ScriptMan::MULTICONNECTTOME_ARRIVAL) == true ||
+	if(ScriptManager->Arrival(curUser, sData, iLen, (uint8_t)(bMulti == false ? ScriptMan::CONNECTTOME_ARRIVAL : ScriptMan::MULTICONNECTTOME_ARRIVAL)) == true ||
 		curUser->iState >= User::STATE_CLOSING) {
 		return;
 	}
